索引库,代码优化

+++ src/main/java/com/bfly/common/FileUtil.java
@@ -5,10 +5,11 @@ import org.apache.commons.io.FileUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.servlet.http.HttpServletRequest;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
+import java.nio.file.Files;
+import java.util.stream.Stream;
 
 /**
  * 文件工具类
@@ -36,7 +37,7 @@ public class FileUtil {
     /**
      * 批量删除文件或文件夹
      *
-     * @param filePath Map对象中的List对象
+     * @param filePath
      * @author 胡礼波
      * 2012-5-9 下午04:38:44
      */
@@ -48,6 +49,29 @@ public class FileUtil {
                     continue;
                 }
                 File file = new File(url);
+                deleteFiles(file);
+            }
+            flag = true;
+        } catch (Exception e) {
+            logger.error("删除文件出错", e);
+        }
+        return flag;
+    }
+
+    /**
+     * 批量删除文件或文件夹
+     *
+     * @param files
+     * @author 胡礼波
+     * 2012-5-9 下午04:38:44
+     */
+    public static boolean deleteFiles(File... files) {
+        boolean flag = false;
+        try {
+            for (File file : files) {
+                if (file == null) {
+                    continue;
+                }
                 FileUtils.deleteQuietly(file);
             }
             flag = true;
@@ -137,5 +161,56 @@ public class FileUtil {
         if (!file.exists()) {
             file.mkdirs();
         }
+    }
+
+    /**
+     * 获得指定路径下的文件名集合--不包括文件夹及文件夹下的文件
+     *
+     * @author andy_hulibo@163.com
+     * @date 2019/8/19 11:36
+     */
+    public static String[] getFileNames(String path) {
+        File mobileFile = new File(path);
+        if (mobileFile.exists()) {
+            Stream<File> stream = Stream.of(mobileFile.listFiles());
+            Stream<String> streamFileName = stream.filter(file -> !file.isDirectory()).map(file -> file.getName());
+            return streamFileName.toArray(String[]::new);
+        }
+        return null;
+    }
+
+    /**
+     * 匹配对应的类型
+     *
+     * @param imgFile
+     * @return
+     * @author 胡礼波
+     * 2014-3-19 上午11:00:47
+     */
+    public static String getImageContentType(File imgFile) {
+        String contentType = null;
+        try {
+            contentType = Files.probeContentType(imgFile.toPath());
+        } catch (IOException e) {
+            logger.error("获得图片类型出错", e);
+        }
+        switch (contentType) {
+            case "image/x-png":
+            case "image/png":
+                contentType = "png";
+                break;
+
+            case "image/bmp":
+                contentType = "bmp";
+                break;
+
+            case "image/gif":
+                contentType = "gif";
+                break;
+
+            default:
+                contentType = "jpg";
+        }
+        return contentType;
     }
 }
